+++ src/com/kaylerrenslow/armaDialogCreator/arma/header/HeaderArray.java
@@ -1,10 +1,12 @@
 package com.kaylerrenslow.armaDialogCreator.arma.header;
 
+import org.jetbrains.annotations.NotNull;
+
 import java.util.List;
 
 /**
  @author Kayler
  @since 03/19/2017 */
 public interface HeaderArray extends HeaderValue {
-	List<HeaderArrayItem> getItems();
+	@NotNull List<HeaderArrayItem> getItems();
 }